Fix in-call notification icon flicker when screen goes on/off
We currently suppress the "in-call" status bar notification if the
InCallScreen is the foreground activity.  This helps reduce clutter in the
status bar, since it's already obvious that you're on a call.

Unfortunately, when the screen is off (like when the prox sensor is
active), the InCallScreen is technically not the foreground activity, so
we *were* enabling the status bar icon in that state.  This was causing a
brief flicker of the status bar icon any time the screen turned on during
a call (like when pulling the phone away from your face.)

The fix is simple: just suppress the notification when the screen is off.
This way the transition from "screen off" to "InCallScreen in foreground"
doesn't cause the notification to change state.

TESTED:

  - While on a call, with in-call UI visible, trigger prox sensor several
    times
    ==> no glitches; icon is never visible.

  - While on a call, bail out to Home screen, then, trigger prox sensor
    several times
    ==> no glitches; icon is steadily visible whenever the screen is on.
        (Specifically, this test shows that this CL has no effect if some
        other app is active.  That's because once you bail out of the
        in-call UI, the updateInCallNotification() method doesn't even get
        called when the screen goes on and off; it's only called for phone
        state transitions, or when the InCallScreen gets paused or
        resumed.  So the notification remains enabled the whole time.)

  - Launched an outgoing call
    ==> no glitches; icon doesn't appear at all (even as a brief flicker)
        during the transition from the dialer to the InCallScreen.

  - Incoming call
    ==> no glitches; icon is visible while ringing (which is WAI),
        disappears cleanly when you answer the call.

Bug: 6021909
Change-Id: I43625f4f69ce75c662c211dadc7292fb14641073